Impl. 1 user - many habits; CRUD controllers & routes for habits.

+++ api/src/controllers/habitController.ts
@@ -5,9 +5,16 @@ import prisma from '../config/prisma';
 
 // Read all habits
 export const getHabits = async (req: Request, res: Response, next: NextFunction) => {
+  /**
+   * - Get the user ID of the current authenticated user
+   * - filter all habits by the auth user's ID
+   * - Return the habits belonging to this particular user.
+   */
   try {
     // res.json(habits);
-    const habits = await prisma.habit.findMany({ include: { completions: true }});
+    const { user_id: userId } = req.user!;
+    const habits = await prisma.habit.findMany({ where: { userId },
+      include: { completions: true }});
     res.json(habits);
   } catch (error) {
     next(error);
@@ -16,9 +23,16 @@ export const getHabits = async (req: Request, res: Response, next: NextFunction)
 
 // Create a new habit
 export const createHabit = async (req: Request, res: Response, next: NextFunction) => {
+  /**
+   * - Get the name from the request body
+   * - Get the user ID of the current authenticated user
+   * - Create a new habit with the name and the user ID
+   * - Return the new habit created.
+   */
   try {
     const { name } = req.body;
-    const newHabit = await prisma.habit.create({ data: { name }});
+    const { user_id: userId} = req.user!;
+    const newHabit = await prisma.habit.create({ data: { name, userId }});
     res.status(201).json(newHabit);
   } catch (error) {
     next(error);
@@ -27,10 +41,18 @@ export const createHabit = async (req: Request, res: Response, next: NextFunctio
 
 // Edit a habit
 export const editHabit = async (req: Request, res: Response, next: NextFunction) => {
+  /**
+   * - Get the id from the request parameters
+   * - Get the user Id of the currently authenticated user.
+   * - Get the name update from the request's body.
+   * - Use the user Id to filter for the user to update
+   * - Return confirmation of update when completed.
+   */
   try {
     const id = parseInt(req.params.id as string, 10);
+    const { user_id: userId } = req.user!;
     const { name } = req.body;
-    await prisma.habit.update({ where: { id }, data: { name }});
+    await prisma.habit.update({ where: { id, userId }, data: { name }});
     res.json({ message: "Updated successfully !"});
   } catch (error) {
     next(error);
@@ -39,9 +61,17 @@ export const editHabit = async (req: Request, res: Response, next: NextFunction)
 
 // Delete a habit
 export const deleteHabit = async (req: Request, res: Response, next: NextFunction) => {
+  /**
+   * - Get the id of habit from the request parameter
+   * - Get the user ID of the currently authenticated user.
+   * - use the user ID to  filter the habit deletion query
+   * - Return a 204 status response to indicate successful deletion.
+   */
+
   try {
     const id = parseInt(req.params.id as string, 10);
-    await prisma.habit.delete({ where: { id }});
+    const { user_id: userId } = req.user!;
+    await prisma.habit.delete({ where: { id, userId }});
     res.status(204).json({});
   } catch (error) {
     next(error);
